backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pytest==8.3.0
pytest-cov==4.1.0
httpx==0.26.0
python-docx==1.1.0
reportlab==4.0.8
python-multipart==0.0.6
pydantic==2.7.0
